Revised framebuffer alignment
Revised framebuffer dimensions based on vce.cpp target buffer
parameters. its expected to write about 262 scanlines into the buffer,
with each scanline pixel writes wrapped at 2048. With this the expected
target buffer would have at least be 264 height, and a width of 2048.

The other PCE cores should also be fixed with this new target height.

This is PR is intended hopefully to remedy framebuffer alignment  issues thats happening in some platforms. Even if it does not fix it, this still appears to be correct expected parameters.

If anyone sees why height for buffer should be 270, then adjust as
needed.